Start in interrupt-load mode
This mode re-loads the audio buffers inside the update() interrupt code, which is less efficient in CPU use, but avoids the issue of code which never calls yield() or delay(). Use AudioPlayWav::enableEventReading(true) to switch to the event-driven scheme: this can be done at any time, even if a sound is playing.